+++ resources/views/admin/observer/consejo_academico.blade.php
@@ -48,16 +48,18 @@
             <i class="eye blue icon"></i>
           </a>
           
-          <a href="{{ route('observer.edit', $observation->id)}}" class="ui tiny icon button" style="display:inline-block !important">
-            <i class="edit blue icon"></i>
-          </a>
-          {{--    
-          {!!Form::open(['route' => ['observer.destroy', $observation->id],
-          'method' => 'DELETE']) !!}
-          <button class="ui tiny icon button">
-            <i class="cancel red icon"></i>
-          </button>
-          --}}
+          @if(Auth::user()->id == 16)   
+            <a href="{{ route('observer.edit', $observation->id)}}" class="ui tiny icon button" style="display:inline-block !important">
+              <i class="edit blue icon"></i>
+            </a>
+            
+            {!!Form::open(['route' => ['observer.destroy', $observation->id],
+            'method' => 'DELETE']) !!}
+            <button class="ui tiny icon button">
+              <i class="cancel red icon"></i>
+            </button>
+          @endif
+         
           {!! Form::close() !!}
           
         </td>
